Add the README testing example as a spec
The example of how to testing Interactors in the README does not work in
the current version of `interactor`. This is a huge barrier to entry for
anyone looking to use the library. We should update the README to
indicate how tests should actually be done.

Note: This commit is just intended to add the spec to show that it
fails. I will add a second commit in a moment to show the method that
I use to test my Interactors.